Alberto's review

+++ src/IProtocolConfig.sol
@@ -3,7 +3,7 @@
 pragma solidity ^0.8.0;
 
 interface IProtocolConfig {
-    function isValidInterestFee(address vault, uint16 interestFee) external view returns (bool);
+    function isValidInterestFee(address vault, uint256 interestFee) external view returns (bool);
 
     function feeConfig(address vault) external view returns (address protocolFeeReceiver, uint256 protocolFeeShare);
