+++ src/interestRateModels/test/IRMZero.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.0;
 import "../BaseIRM.sol";
 
 contract IRMZero is BaseIRM {
-    function computeInterestRateImpl(address, address, uint32) internal pure override returns (uint72) {
+    function computeInterestRateImpl(address, address, uint32) internal pure override returns (uint256) {
         return 0;
     }
 }
